create a stim Base class
