+++ LWSStimuli/LWSStimulus.py
@@ -2,25 +2,17 @@ import os
 import numpy as np
 import matplotlib.pyplot as plt
 
-from LWSStimuli.LWSStimulusTypeEnum import LWSStimulusTypeEnum, identify_stimulus_type
+from LWSStimulusBase import LWSStimulusBase
+from LWSStimuli.LWSStimulusTypeEnum import LWSStimulusTypeEnum
 
 
-class LWSStimulus:
+class LWSStimulus(LWSStimulusBase):
 
-    def __init__(self, stim_id: int, stim_type, directory: str):
-        self.__stim_id = stim_id
-        self.__stim_type = identify_stimulus_type(stim_type)
-        if not os.path.isdir(directory):
-            raise NotADirectoryError(f"Directory {directory} does not exist.")
-        self.__stimulus_directory = directory
-
-    @property
-    def stim_id(self) -> int:
-        return self.__stim_id
-
-    @property
-    def stim_type(self) -> LWSStimulusTypeEnum:
-        return self.__stim_type
+    def __init__(self, stim_id: int, stim_type, super_dir: str):
+        super().__init__(stim_id, stim_type)
+        if not os.path.isdir(super_dir):
+            raise NotADirectoryError(f"Directory {super_dir} does not exist.")
+        self.__super_dir = super_dir
 
     @property
     def image_array_path(self) -> str:
@@ -42,14 +34,8 @@ class LWSStimulus:
     def __path_to_file(self, format: str) -> str:
         subdir = self.stim_type.name.lower()
         filename = f"image_{self.__stim_id}.{format}"
-        fullpath = os.path.join(self.__stimulus_directory, subdir, filename)
+        fullpath = os.path.join(self.__super_dir, subdir, filename)
         if not os.path.isfile(fullpath):
             raise FileNotFoundError(f"File {fullpath} does not exist.")
         return fullpath
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}_{self.stim_type.name.upper()}{self.stim_id}"
-
-    def __str__(self) -> str:
-        return f"{self.__class__.__name__}_{self.stim_type.name.upper()}{self.stim_id}"
